+++ src/Modulus.Cli/Program.cs
@@ -19,6 +19,7 @@ rootCommand.Subcommands.Add(AddEventCommand.Create(fileSystem, consoleOutput));
 rootCommand.Subcommands.Add(AddConsumerCommand.Create(fileSystem, consoleOutput));
 rootCommand.Subcommands.Add(ListModulesCommand.Create(fileSystem, consoleOutput));
 rootCommand.Subcommands.Add(OutboxCommand.Create(fileSystem, consoleOutput));
+rootCommand.Subcommands.Add(DlqCommand.Create(fileSystem, consoleOutput));
 rootCommand.Subcommands.Add(DoctorCommand.Create(fileSystem, consoleOutput));
 rootCommand.Subcommands.Add(UpgradeCommand.Create(fileSystem, consoleOutput));
 rootCommand.Subcommands.Add(VersionCommand.Create(consoleOutput));